Add first attempt at MCU and MAC/PHY schematics
MCU is standard ATMEAG328PB with SPI connection to the ENC28J60 Ethernet
PHY/MAC controller and an ICSP header for programming. The ATMEGA328PB
was chosen because it is a familiar platform for Arduino users, with
additional serial interfaces to what the ATMEGA328P has.

Wiring of the ENC28J60 is based on the schematic of the MOD-ENC28J60
Ethernet Development Board from Olimex.

The Amphenol RJMG1BD3B8K1ANR 8P8C RJ45 Connector was chosen because it
is a part of the default KiCad library and includes the center-taped
pulse transformers required by the ENC28J60.

Makes progress towards DS-9