Replace MCU with more effective alternative
The ATMEGA328PB raised concerns about its ability to process input data
at an acceptable speed, and so it has beed replaced with the
ATSAME51J20A, the same MCU used in the Helix expansion boards. This will
also increase cross-compatibility with components and will decrease the
number of unique components used in Avionics.